feat: switch to use the wording compressed CMAF instead of LOC
